TNO-2607 Fix reports
